requirements.txt
uvicorn
fastapi
jinja2
python-multipart
pillow
bcrypt